+++ src/main/java/nextstep/dto/console/response/ReservationConsoleResponse.java
@@ -1,4 +1,4 @@
-package nextstep.dto.response;
+package nextstep.dto.console.response;
 
 import nextstep.domain.Reservation;
